+++ config/application.rb
@@ -15,5 +15,6 @@ module CardswiprNew
     # Application configuration can go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded after loading
     # the framework and any gems in your application.
+    config.eager_load_paths += ["#{config.root}/lib"]
   end
 end