+++ db/migrate/20190624160724_change_check_in_to_datetime_in_attendances.rb
@@ -1,0 +1,6 @@
+class ChangeCheckInToDatetimeInAttendances < ActiveRecord::Migration[5.2]
+  def change
+  	remove_column :attendances, :check_in
+  	add_column :attendances, :check_in, :datetime
+  end
+end