+++ spec/support/person_api_helpers.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+require './lib/people_hub/person_request'
+# Person API helper function
+def file_to_response(path, code_num)
+  unless path.nil?
+    file = File.read(path)
+    format = File.extname(path).from(1).downcase.to_sym
+    parser = HTTParty::Parser.new(file, format)
+    data = parser.parse
+  end
+  instance_double('response',
+                  response: instance_double('child_response'),
+                  code: code_num, parsed_response: data)
+end